+++ src/frontend/apps/impress/src/core/config/ConfigProvider.tsx
@@ -1,0 +1,9 @@
+import { PropsWithChildren } from 'react';
+
+import { useConfig } from './api/useConfig';
+
+export const ConfigProvider = ({ children }: PropsWithChildren) => {
+  useConfig();
+
+  return children;
+};