+++ src/frontend/apps/impress/src/core/config/api/useConfig.tsx
@@ -1,0 +1,33 @@
+import { useQuery } from '@tanstack/react-query';
+
+import { APIError, errorCauses, fetchAPI } from '@/api';
+
+interface ConfigResponse {
+  SENTRY_DSN: string;
+  COLLABORATION_SERVER_URL: string;
+  ENVIRONMENT: string;
+  FRONTEND_THEME: string;
+  LANGUAGES: [string, string][];
+  LANGUAGE_CODE: string;
+  MEDIA_BASE_URL: string;
+}
+
+export const getConfig = async (): Promise<ConfigResponse> => {
+  const response = await fetchAPI(`config/`);
+
+  if (!response.ok) {
+    throw new APIError('Failed to get the doc', await errorCauses(response));
+  }
+
+  return response.json() as Promise<ConfigResponse>;
+};
+
+export const KEY_CONFIG = 'config';
+
+export function useConfig() {
+  return useQuery<ConfigResponse, APIError, ConfigResponse>({
+    queryKey: [KEY_CONFIG],
+    queryFn: () => getConfig(),
+    staleTime: Infinity,
+  });
+}